22.01.12

+++ src/main/webapp/resources/js/reply.js
@@ -24,104 +24,121 @@ $(document).ready(function() {
 	showList(1);	// showList 함수 호출
 	
 	
-	// p440
-	var pageNum = 1;
-	var replyPageFooter = $(".reply-footer");
-	
-	function showReplyPage(replyCnt) {
-		
-		var endNum = Math.ceil(pageNum/10.0) * 10;
-		var startNum = endNum - 9;
-		
-		var prev = startNum != 1;
-		var next = false;
-		
-		if (endNum * 10 >= replyCnt) {
-			endNum = Math.ceil(replyCnt/10.0);
-		}
-		
-		if (endNum * 10 < replyCnt) {
-			next = true;
-		}
-		
-		var str = "<ul class='pagenation' style='display: flex;justify-content:center;'>";
-		
-		if (prev) {
-			str += "<li><a class='pagenation__arr pagenation__arr--prev' href ='" + (startNum - 1) + "'>Previous</a></li>";
-		}
-		
-		for(var i = startNum; i <= endNum; i++) {
-			var active = pageNum == i? "active" : "";
-			
-			str += "<li style='margin: 0 5px;'><a class='" + active + "' href='" + i + "' style='color: #777;'>" + i + "</a></li>";
-		}
-		
-		if (next) {
-			str += "<li><a class='pagenation__arr pagenation__arr--next' href='" + (endNum + 1) + " '>Next</a></li>";
-		}
-		
-		str += "</ul></div>";
-		
-		// console.log(str)
-		replyPageFooter.html(str);
-	}
-	
-	// showList 함수 선언 시작
+//	var pageNum = 1;
+//	var replyPageFooter = $(".reply-footer");
+//	
+//	function showReplyPage(replyCnt) {
+//		
+//		var endNum = Math.ceil(pageNum/10.0) * 10;
+//		var startNum = endNum - 9;
+//		
+//		var prev = startNum != 1;
+//		var next = false;
+//		
+//		if (endNum * 10 >= replyCnt) {
+//			endNum = Math.ceil(replyCnt/10.0);
+//		}
+//		
+//		if (endNum * 10 < replyCnt) {
+//			next = true;
+//		}
+//		
+//		var str = "<ul class='pagenation' style='display: flex;justify-content:center;'>";
+//		
+//		if (prev) {
+//			str += "<li><a class='pagenation__arr pagenation__arr--prev' href ='" + (startNum - 1) + "'>Previous</a></li>";
+//		}
+//		
+//		for(var i = startNum; i <= endNum; i++) {
+//			var active = pageNum == i? "active" : "";
+//			
+//			str += "<li style='margin: 0 5px;'><a class='" + active + "' href='" + i + "' style='color: #777;'>" + i + "</a></li>";
+//		}
+//		
+//		if (next) {
+//			str += "<li><a class='pagenation__arr pagenation__arr--next' href='" + (endNum + 1) + " '>Next</a></li>";
+//		}
+//		
+//		str += "</ul></div>";
+//		
+//		// console.log(str)
+//		replyPageFooter.html(str);
+//	}
+//	
 	function showList(page) {
 		// getList 함수 호출 시작
 		
 		//	alert("aaa");
 		
-		replyService.getList({bno:bnoValue, page:page||1}, function(replyCnt, list) {
+		replyService.getList({bno:bnoValue, page:page||1}, function(list) {
 			
-			
-			var rereplyT = $("#rereply");				
-			var registerBtn = $("#registertBtn");
-			
-		
-							//'{ ~ }'은 JSON type			'function(list)'는 callback 함수
-				if(page == -1) {
-					pageNum = Math.ceil(replyCnt/10.0);
-					showList(pageNum);
-					return;
-				}
-							
+				console.log(list.length-1)
 							
 				var str = "";
 				
-				// select 된 결과가 없으면
 				if(list == null || list.length==0) {
 					replyUL.html("");	
 					return;
 				}
-				
 				console.log(list);
-				
 				for(var i = 0, len=list.length || 0; i < len; i++) {
 					str+="<li class='reply_view'>";
 					str+="<div style='display:flex;margin-left:"+40*list[i].redepth+"px;'>";
-					str+="<div style='margin-left: 30px;display:flex;'><span style='margin:0 15px;'>"+list[i].replyer+"</span><i class='reply'>"+list[i].reply+"</i></div>";
-                    str+="<div style='display:flex;margin-left: 10px;'><div style='margin-right:20px'>"+replyService.displayTime(list[i].replyDate)+"</div>";
-					str+="<div id='ModBtn'>수정</div>";
+					str+="<div style='margin-left: 30px;display:flex;'>";
+					str+="<span class='replyer' style='margin:0 15px;'>"+list[i].replyer+"</span>";
+					str+="<i class='reply'>"+list[i].reply+"</i></div>";
+					if(list[i].replyer!="삭제된 댓글 입니다."){
 					
-					if(list[i].redepth==0){
-					str+="<div id='reChat' style='cursor:pointer'>&nbsp&nbsp답글</div>";
+	                    str+="<div class='right' style='display:flex;margin-left: 10px;'><div style='margin-right:20px'>"+replyService.displayTime(list[i].replyDate)+"</div>";
+						str+="<div id='ModBtn'>수정</div>";
+						
+						if(list[i].redepth==0){
+						str+="<div id='reChat' style='cursor:pointer'>&nbsp&nbsp답글</div>";
+						}
+						
+						str+="<div id='RemoveBtn'>&nbsp&nbsp삭제</div>";
+						str+="<span class='rnoColumn'>"+list[i].rno+"</span><i style='display:none'>"+list[i].redepth+"</i></div>";
+	                    str+="<span class='blindBtn'>수정</span><i class='blindBtn closer' style='margin-left: 8px;'>취소</i>";
+						str+="</div>";	
+	                    str+="<p class='rcIn' style='display:flex;'><span></span><i style='display:none;margin-top: 9px;'><button id='reRegisterBtn' class='btn' style='margin-left: 15px;'>등록</button><button class='btn close' style='margin-left: 8px;'>취소</button></i></p>";
+					} else{
+						if(list[i]==list[list.length-1]){
+							var tRno=list[i].rno;
+							replyService.remove( 
+								tRno,				 
+								function(count){ 
+								console.log(count);
+									if(count === "success"){
+										showList(1);
+									}
+								},				 
+								function(error){	 
+									alert("error.....");
+								}					 
+							)
+						} else if(list[i+1].redepth!=1){
+							var tRno=list[i].rno;
+							replyService.remove( 
+								tRno,				 
+								function(count){ 
+								console.log(count);
+									if(count === "success"){
+										showList(1);
+									}
+								},				 
+								function(error){	 
+									alert("error.....");
+								}					 
+							)
+						}
 					}
-					
-					str+="<div id='RemoveBtn'>&nbsp&nbsp삭제</div>";
-					str+="<span class='rnoColumn'>"+list[i].rno+"</span></div>";
-                    str+="<span class='blindBtn'>수정</span><i class='blindBtn closer' style='margin-left: 8px;'>취소</i>";
-					str+="</div>";	
-                    str+="<p class='rcIn' style='display:flex;'><span></span><i style='display:none;margin-top: 9px;'><button id='reRegisterBtn' class='btn' style='margin-left: 15px;'>등록</button><button class='btn close' style='margin-left: 8px;'>취소</button></i></p>";
                     str+="</li>";
-					if($("#reChat").find('span').text()==1){$("#reChat").css('display','none');}else{$("#reChat").css('display','block');}
 				} 
 				
 				replyUL.html(str);
 				
-				showReplyPage(replyCnt);
-			}) // getList 함수 호출 끝
-	}	// showList 함수 선언 끝
+			}) 
+	} 
 	
 	
 	var replyT = $("#reply");				
@@ -134,26 +151,17 @@ $(document).ready(function() {
 				reply : replyT.val(),
 				bno : bnoValue
 			};
-		replyService.add(reply, function(result) {
-			if (confirm("등록 하시겠습니까?")){
-			replyT.val("");
-			showList(1);
-			} else{replyT.val("");};
-			}
-			)
-			
+		if (confirm("등록 하시겠습니까?")){
+			replyService.add(reply, function(result) {
+					replyT.val("");
+					showList(1);
+					}
+					)
+			} else{replyT.val("");};			
 		})
 		$('.closer').click(function(){if(confirm("취소 하시겠습니까?")){$(this).parent().parent().css('display','none')}})
 
-//		$(".chat").on("click", "#reChat", function(e) {
-////			e.preventDefault();
-//			var rno=$(this).siblings("span").html();
-//			var reply = $(this).parent().siblings("div").find("i").html();
-//			console.log("aaa "+rno+" bbb "+reply);
-//			var paReply ={rno:rno, reply: reply ,reparent : rno}
-//			replyService.update(paReply, function(result) {})
-//		})
-//		
+
 		$(".chat").on("click", "#reChat", function(e) {
 			e.preventDefault();
 			var rno=$(this).siblings("span").html();
@@ -175,7 +183,7 @@ $(document).ready(function() {
 						reparent : rno,
 						redepth : 1
 					};
-				
+				if(reVal.val()=="" || reVal.val().length==0){alert("대댓글을 입력해주세요."); return false;}
 					if (confirm("등록 하시겠습니까?")){
 						replyService.add(reply, function(result) {reVal.val("");showList(1);})
 						
@@ -186,21 +194,28 @@ $(document).ready(function() {
 		
 		$(".chat").on("click", "#RemoveBtn", function(e) {
 					var rno=$(this).siblings("span").html();
+					var dept=$(this).siblings("i").html();
+					var par=$(this).parent().parent().parent().next("li").find("div").css('margin-left')
+					console.log(par!="0px");
 				if (confirm("삭제 하시겠습니까?")){
-					replyService.remove( 
-							rno,				 
-							function(count){ 
-							console.log(count);
-							
-								if(count === "success"){
-									showList(pageNum);
-									
-								}
-							},				 
-							function(error){	 
-								alert("error.....");
-							}					 
-						)
+					 if (dept==1 || par!="40px"){
+						replyService.remove( 
+								rno,				 
+								function(count){ 
+								console.log(count);
+									if(count === "success"){
+										showList(1);
+									}
+								},				 
+								function(error){	 
+									alert("error.....");
+								}					 
+							)
+						} else if(dept==0){
+						replyService.delUpdate(rno, function(result) {
+							showList(1);
+						});
+					}
 					}
 						return false;
 		})
@@ -220,7 +235,7 @@ $(document).ready(function() {
 			par.contents().unwrap().wrap("<input style='width: 185px;margin: 0 15px;' type='text' class='modInput' name=reply value='"+reply+"'>");
 			cBtn.click(function(){
 				if (confirm("취소 하시겠습니까?")){
-					showList(pageNum);
+					showList(1);
 				}
 			})
 			bBtn.on("click", function(e) {
@@ -229,7 +244,7 @@ $(document).ready(function() {
 			if(ModReplypar.length==0 || ModReplypar==" "){alert("댓글을 입력해주세요."); return false;}
 				if (confirm("수정 하시겠습니까?")){
 			replyService.update(reply, function(result) {
-				showList(pageNum);
+				showList(1);
 			});
 			}
 //			else{$(".modInput").val("")}
@@ -239,16 +254,16 @@ $(document).ready(function() {
 		// 댓글 삭제 처리
 
 	// 댓글의 페이징 버튼을 클릭하면
-	replyPageFooter.on("click","li a",function(e){
-		e.preventDefault();
-		console.log("page click");
-		var targetPageNum=$(this).attr("href");
-		
-		console.log("targetPageNum : "+targetPageNum);
-		pageNum = targetPageNum;
-		showList(pageNum);
-		modal.find("input").val("");
-	})
+//	replyPageFooter.on("click","li a",function(e){
+//		e.preventDefault();
+//		console.log("page click");
+//		var targetPageNum=$(this).attr("href");
+//		
+//		console.log("targetPageNum : "+targetPageNum);
+//		pageNum = targetPageNum;
+//		showList(pageNum);
+//		modal.find("input").val("");
+//	})
 	})
 
 
@@ -280,28 +295,6 @@ var replyService = (function() {
 		})													 
 	}		
 	
-	function readd(reply, callback, error) {				 
-		console.log("reply......")
-		
-		
-		$.ajax({											 
-			type: 'post',
-			url : '/rereplies/new',	 
-			data : JSON.stringify(reply),
-			contentType : "application/json; charset=utf-8",
-			success : function(result, status, xhr) {
-				if (callback) {
-					callback(result);
-				}
-			},
-			error : function(xhr, status, er) {
-				if (error) {
-					error(er);
-				}
-			}
-		})													 
-	}													 
-	
 	function getList(param, callback, error) {			 
 			var bno = param.bno;
 			var page = param.page || 1;
@@ -309,7 +302,7 @@ var replyService = (function() {
 			$.getJSON("/replies/pages/" + bno + "/" + page + ".json", 
 				function(data) {
 					if (callback) {
-						callback(data.replyCnt, data.list);
+						callback(data);
 					}
 			}) 
 			.fail(function(xhr, status, err) {
@@ -340,6 +333,27 @@ var replyService = (function() {
 	}													 
 	
 	
+	function delUpdate(rno, callback, error) {				 
+		console.log("reply......")
+		
+		
+		$.ajax({											 
+			type: 'post',
+			url: "/replies/" + rno,
+			data : JSON.stringify(reply),
+			contentType : "application/json; charset=utf-8",
+			success : function(result, status, xhr) {
+				if (callback) {
+					callback(result);
+				}
+			},
+			error : function(xhr, status, er) {
+				if (error) {
+					error(er);
+				}
+			}
+		})													 
+	}				
 	
 	// update 함수
 	function update(reply, callback, error) {			 
@@ -415,7 +429,7 @@ var replyService = (function() {
 	
 	return {
 		add:add, 
-		readd:readd, 
+		delUpdate:delUpdate, 
 		getList:getList,
 		remove:remove,
 		update:update,
